split first pikachu file, and my first time splitting ever
